refactoring + upload cmd changed
