+++ pkg/gomod/parser.go
